remove old radio and input and delete unused files
