--- apps/betterangels/src/app/(tabs)/appointment.tsx
@@ -1,5 +1,0 @@
-import { SafeAreaView } from 'react-native';
-
-export default function AppointmentScreen() {
-  return <SafeAreaView style={{ flex: 1 }}></SafeAreaView>;
-}
