--- apps/betterangels/src/app/(tabs)/calendar.tsx
@@ -1,10 +1,0 @@
-import { MainScrollContainer } from '@monorepo/expo/betterangels';
-import { TextRegular } from '@monorepo/expo/shared/ui-components';
-
-export default function CalendarScreen() {
-  return (
-    <MainScrollContainer>
-      <TextRegular>Calendar Screen</TextRegular>
-    </MainScrollContainer>
-  );
-}
