--- apps/betterangels/src/app/(tabs)/drawerPlaceholder.tsx
@@ -1,3 +1,0 @@
-export default function DrawerPlaceholderScreen() {
-  return null;
-}
