--- apps/betterangels/src/app/(tabs)/map.tsx
@@ -1,10 +1,0 @@
-import { MainScrollContainer } from '@monorepo/expo/betterangels';
-import { TextRegular } from '@monorepo/expo/shared/ui-components';
-
-export default function MapScreen() {
-  return (
-    <MainScrollContainer>
-      <TextRegular>Map</TextRegular>
-    </MainScrollContainer>
-  );
-}
