--- apps/betterangels/src/app/team/_layout.tsx
@@ -1,11 +1,0 @@
-import { useUser } from '@monorepo/expo/betterangels';
-import { Redirect, Slot } from 'expo-router';
-
-export default function TeamLayout() {
-  const { user } = useUser();
-
-  if (!user) {
-    return <Redirect href="/auth" />;
-  }
-  return <Slot />;
-}
