+++ libs/expo/betterangels/src/lib/ui-components/index.ts
@@ -1,4 +1,3 @@
-export { default as AdaAccommodationPicker } from './AdaAccommodationPicker';
 export { default as AddressField } from './AddressField';
 export { AppUpdatePrompt } from './AppUpdatePrompt';
 export { default as AuthContainer } from './AuthContainer';
@@ -12,24 +11,17 @@ export {
 export { default as ConsentModal } from './ConsentModal';
 export { default as DocumentModal } from './DocumentModal';
 export { default as ErrorCrashView } from './ErrorCrashView/ErrorCrashView';
-export { default as EyeColorPicker } from './EyeColorPicker';
 export { default as FeatureFlagControlled } from './FeatureFlagControlled/FeatureFlagControlled';
 export { FileThumbnail } from './FileThumbnail/FileThumbnail';
-export { default as GenderPicker } from './GenderPicker';
-export { default as HairColorPicker } from './HairColorPicker';
 export { default as Header } from './Header';
 export { default as LoginForm } from './LoginForm';
 export { default as MainContainer } from './MainContainer';
 export { default as MainPlusModal } from './MainPlusModal';
 export { default as MainScrollContainer } from './MainScrollContainer';
-export { default as MaritalStatusPicker } from './MaritalStatusPicker';
 export { default as Modal } from './Modal';
 export * from './Nav';
 export { default as NavModal } from './NavModal';
 export { default as NoteCard } from './NoteCard';
-export { default as PreferredCommunicationPicker } from './PreferredCommunicationPicker';
-export { default as PronounPicker } from './PronounPicker';
-export { default as RacePicker } from './RacePicker';
 export { default as RequestedProvidedServices } from './RequestedProvidedServices';
 export { default as OtherCategory } from './RequestedProvidedServices/OtherCategory';
 export { AppleSignIn, GoogleSignIn } from './SocialLogin';
