+++ libs/expo/shared/ui-components/src/lib/Input/InputClearIcon.tsx
