--- libs/expo/shared/ui-components/src/lib/Input_V2/index.ts
@@ -1,1 +1,0 @@
-export { Input as default } from './Input';
