--- libs/expo/shared/ui-components/src/lib/Radio_V2/index.ts
@@ -1,1 +1,0 @@
-export { Radio as default } from './Radio';
