+++ libs/expo/shared/ui-components/src/lib/index.ts
@@ -39,7 +39,6 @@ export { default as FormFieldLabel } from './FormFieldLabel';
 export { default as IconButton } from './IconButton';
 export { default as ImageViewer } from './ImageViewer';
 export { default as Input } from './Input';
-export { default as Input_V2 } from './Input_V2';
 export { default as KeyboardAwareScrollView } from './KeyboardAwareScrollView';
 export {
   Length,
